Fix pass_fptr_to_wt 64-bit return truncation across the grate dispatch chain
Grate-interposed calls returning a full 64-bit scalar (double, int64_t,
pointer) were silently corrupted: the wasm-level portal in linker.rs
already used the callee's real function type, but the value flowing
through dispatch_lib_call -> _call_grate_func -> the runtime trampoline
-> pass_fptr_to_wt was narrowed to i32/int at every hop, then
reinterpreted as a full 64-bit double at the portal (sign-extending a
truncated 32-bit half into a bogus bit pattern). Confirmed root cause
via the glibc libm test suite: nextafter's real value got wiped by
check_ulp()'s startup sanity check, collapsing the entire double tier
under interposition.

Widen the whole chain to i64/int64_t, leaving the unrelated plain
make_syscall/RawCallFunc path (regular Linux syscalls) at i32:
- lind_marshal.h: LIND_DEFINE_MARSHAL_HANDLER macro
- gen_grate.py: pass_fptr_to_wt in both GRATE_TEMPLATE and
  FREESTANDING_TEMPLATE, plus the regenerated libm_full_grate.c
- threei.rs/threei_const.rs: GrateTrampolineFn, _call_grate_func,
  dispatch_lib_call, GRATE_OK/GRATE_ERR (make_syscall's own two direct
  call sites explicitly truncate back to i32 at the boundary)
- lind-3i: PassFptrTyped, GrateWorker::run, GrateHandler::submit*
- lind-boot: grate_callback_trampoline, submit_grate_request (the live
  path lind_run actually uses)
- linker.rs: the portal itself — I64/F64 now pass the value straight
  through instead of the old `ret as u64` sign-extension bug; I32/F32
  take the low 32 bits explicitly
- commands/run.rs: mirrored for consistency (wasmtime-cli dev path,
  not used by lind_run)

Verified via the full 411-test glibc libm suite (dynamically linked,
libm_full_grate.cwasm): 22 tests moved CRASH -> REAL_FAIL/FLAG_ONLY,
zero regressions, CLEAN count unchanged (102/102). nextafter and cos
now match baseline exactly. Remaining crashes are the separate,
already-known long-double (>64-bit) marshalling gap.

+++ src/lind-boot/src/lind_wasmtime/trampoline.rs
@@ -26,7 +26,7 @@ pub extern "C" fn grate_callback_trampoline(
     arg5cageid: u64,
     arg6: u64,
     arg6cageid: u64,
-) -> i32 {
+) -> i64 {
     // Form the grate request with the provided arguments and the handler address
     let req = GrateRequest {
         handler_addr: in_grate_fn_ptr_u64,
